requirements.txt
#langchain==0.1.0
#langchain-community>=0.0.10
#python-dotenv==1.0.0
#streamlit>=1.28.0 
#langchain-openai==0.1.0

langchain
langchain-community
python-dotenv
streamlit
langchain-openai
openai
pypdf
chromadb
langchain_chroma
faiss-cpu